refactor

+++ packages/core/src/EffectHook.js
@@ -8,36 +8,36 @@ export const EFFECT_LAYOUT = 1;
 /**
  *
  * @param {number} tag
- * @param {function} callback
+ * @param {function} mount
  * @param {[]|undefined} deps
  * @constructor
  */
-export function EffectHook(tag, callback, deps) {
+export function EffectHook(tag) {
     this.tag_ = tag;
-    this.callback_ = callback;
-    this.deps_ = deps;
+    this.mount_ = undefined;
+    this.deps_ = undefined;
     this.destroy_ = undefined;
-    this.lastDeps_ = undefined;
+    this.prevDeps_ = undefined;
     this.next_ = null;
 }
 
-export let useEffect = (callback, deps) => {
-    return _useEffectImpl(EFFECT_NORMAL, callback, deps);
+export let useEffect = (mount, deps) => {
+    return _useEffectImpl(EFFECT_NORMAL, mount, deps);
 };
 
-export let useLayoutEffect = (callback, deps) => {
-    return _useEffectImpl(EFFECT_LAYOUT, callback, deps);
+export let useLayoutEffect = (mount, deps) => {
+    return _useEffectImpl(EFFECT_LAYOUT, mount, deps);
 };
 
-let _useEffectImpl = (tag, callback, deps) => {
+let _useEffectImpl = (tag, mount, deps) => {
     return resolveCurrentEffectHook(
         (currentVNode) => {
-            return new EffectHook(tag, callback, deps);
+            return new EffectHook(tag);
         },
         (currentHook) => {
             warnIfDepsSizeChangedOnDEV(deps, currentHook.deps_);
             
-            currentHook.callback_ = callback;
+            currentHook.mount_ = mount;
             currentHook.deps_ = deps;
         }
     );
@@ -52,10 +52,16 @@ let _useEffectImpl = (tag, callback, deps) => {
 export let mountEffects = (effectTag, vnode, isNewlyMounted) => {
     let hook = vnode.effectHook_;
     while (hook !== null) {
-        if (hook.tag_ === effectTag) {
-            if (isNewlyMounted || depsMismatch(hook.deps_, hook.lastDeps_)) {
+        if (hook.tag_ === effectTag && hook.mount_ !== undefined) {
+            if (isNewlyMounted || depsMismatch(hook.deps_, hook.prevDeps_)) {
+                // Save the previous deps
+                hook.prevDeps_ = hook.deps_;
+                hook.deps_ = undefined;
+
+                // Run the effect mount
                 try {
-                    _mountEffect(hook);
+                    hook.destroy_ = hook.mount_();
+                    hook.mount_ = undefined;
                 } catch (error) {
                     catchError(error, vnode);
                 }
@@ -73,29 +79,16 @@ export let mountEffects = (effectTag, vnode, isNewlyMounted) => {
 export let destroyEffects = (effectTag, vnode, isUnmounted) => {
     let hook = vnode.effectHook_;
     while (hook !== null) {
-        if (hook.tag_ === effectTag) {
-            if (hook.destroy_ !== undefined) {
-                if (isUnmounted || depsMismatch(hook.deps_, hook.lastDeps_)) {
-                    try {
-                        hook.destroy_();
-                    } catch (error) {
-                        catchError(error, vnode);
-                    }
+        if (hook.tag_ === effectTag && hook.destroy_ !== undefined) {
+            if (isUnmounted || depsMismatch(hook.deps_, hook.prevDeps_)) {
+                try {
+                    hook.destroy_();
+                    hook.destroy_ = undefined;
+                } catch (error) {
+                    catchError(error, vnode);
                 }
             }
         }
         hook = hook.next_;
     }
-};
-
-/**
- *
- * @param {EffectHook} hook
- */
-let _mountEffect = (hook) => {
-    // Save the last deps for the next time
-    hook.lastDeps_ = hook.deps_;
-    
-    // Run the effect callback
-    hook.destroy_ = hook.callback_();
 };
